[CWS] move common process code outside of linux specific files
